Add user id in post table

+++ src/api/post/sql/deletePost.sql
@@ -1,3 +1,4 @@
 DELETE FROM post
 WHERE id = $1
+  AND user_id = $2
 RETURNING id
